🔧 tx SetCollectionMaxSupply
